% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitch.R
\docType{data}
\name{pitch_meters}
\alias{pitch_meters}
\alias{pitch_yards}
\title{Standard soccer pitch dimensions}
\format{
A named list of length 12.

An object of class \code{list} of length 12.
}
\usage{
pitch_meters

pitch_yards
}
\description{
Named lists holding every dimension needed to draw a pitch. Pass one of these
(or a list with the same names) to the \code{pitch_dimensions} argument of
\code{\link{pitch}}. \code{pitch_meters} holds the metric dimensions
(the default) and \code{pitch_yards} the equivalent imperial dimensions.
}
\details{
Each list contains:
\describe{
  \item{length}{Length of the pitch (goal line to goal line).}
  \item{width}{Width of the pitch (touchline to touchline).}
  \item{penalty_area_depth}{Depth of the penalty area, from the goal line.}
  \item{penalty_area_width}{Width of the penalty area.}
  \item{goal_area_depth}{Depth of the goal area (the small box).}
  \item{goal_area_width}{Width of the goal area.}
  \item{penalty_spot_dist}{Distance of the penalty spot from the goal line.}
  \item{penalty_arc_radius}{Radius of the penalty arc (the "D").}
  \item{centre_circle_radius}{Radius of the centre circle.}
  \item{corner_arc_radius}{Radius of the corner arcs.}
  \item{goal_width}{Width of the goal mouth.}
  \item{goal_depth}{Depth the goal mouth extends behind the goal line.}
}
}
\keyword{datasets}
